+++ codex-rs/tui/src/entertainment/generator.rs
@@ -1,4 +1,3 @@
-use std::collections::VecDeque;
 use std::sync::Arc;
 
 use codex_core::ThreadManager;
@@ -6,78 +5,13 @@ use codex_core::config::Config;
 use codex_core::protocol::EventMsg;
 use codex_core::protocol::Op;
 use codex_protocol::user_input::UserInput;
-use ratatui::text::Line;
 use serde::Deserialize;
 use serde_json::Value;
 use tracing::info;
 
-use crate::app_event::AppEvent;
-use crate::app_event_sender::AppEventSender;
-use crate::history_cell::HistoryCell;
-
-const PROMPT_TEMPLATE: &str = include_str!("entertainment_prompt.md");
-const HISTORY_LIMIT: usize = 10;
-
-#[derive(Debug)]
-pub(crate) struct EntertainmentTextManager {
-    enabled: bool,
-    history: VecDeque<String>,
-}
-
 #[derive(Debug, Deserialize)]
 struct EntertainmentTextOutput {
     texts: Vec<String>,
-}
-
-impl EntertainmentTextManager {
-    pub(crate) fn new(enabled: bool) -> Self {
-        Self {
-            enabled,
-            history: VecDeque::new(),
-        }
-    }
-
-    pub(crate) fn record_history_cell(&mut self, cell: &dyn HistoryCell) {
-        let lines = cell.transcript_lines(u16::MAX);
-        let text = render_lines(&lines).join("\n");
-        let text = text.trim();
-        if text.is_empty() {
-            return;
-        }
-        self.history.push_back(text.to_string());
-        while self.history.len() > HISTORY_LIMIT {
-            self.history.pop_front();
-        }
-    }
-
-    pub(crate) fn request_generation(&self, app_event_tx: &AppEventSender) {
-        if !self.enabled {
-            return;
-        }
-        let prompt = self.build_prompt();
-        info!(
-            history_len = self.history.len(),
-            prompt_len = prompt.len(),
-            "requesting entertainment text generation"
-        );
-        app_event_tx.send(AppEvent::GenerateEntertainmentTexts { prompt });
-    }
-
-    fn build_prompt(&self) -> String {
-        let history = if self.history.is_empty() {
-            "- (no recent history)".to_string()
-        } else {
-            let mut out = String::new();
-            for entry in &self.history {
-                out.push_str("- ");
-                out.push_str(entry);
-                out.push('\n');
-            }
-            out.trim_end().to_string()
-        };
-
-        PROMPT_TEMPLATE.replace("{{INSERT_CONTEXT_HERE}}", &history)
-    }
 }
 
 pub(crate) async fn generate_entertainment_texts(
@@ -90,7 +24,7 @@ pub(crate) async fn generate_entertainment_texts(
         "starting entertainment text generation thread"
     );
     let mut config = config;
-    config.model = Some("gpt-4.1-nano".to_string());
+    config.model = Some("gpt-5-nano".to_string());
     let new_thread = server.start_thread(config).await?;
     let schema = entertainment_output_schema();
     let input = vec![UserInput::Text { text: prompt }];
@@ -173,16 +107,4 @@ fn entertainment_output_schema() -> Value {
         "required": ["texts"],
         "additionalProperties": false
     })
-}
-
-fn render_lines(lines: &[Line<'static>]) -> Vec<String> {
-    lines
-        .iter()
-        .map(|line| {
-            line.spans
-                .iter()
-                .map(|span| span.content.as_ref())
-                .collect::<String>()
-        })
-        .collect()
 }
