+++ codex-rs/tui/src/entertainment/mod.rs
@@ -1,0 +1,11 @@
+pub(crate) mod arc_store;
+pub(crate) mod controller;
+pub(crate) mod generator;
+pub(crate) mod shimmer_text;
+pub mod test_support;
+
+pub(crate) use arc_store::EntertainmentArcStore;
+pub(crate) use controller::EntertainmentController;
+pub(crate) use generator::generate_entertainment_texts;
+pub(crate) use shimmer_text::ShimmerStep;
+pub(crate) use shimmer_text::ShimmerText;
