+++ codex-rs/tui/src/entertainment/prompt.md
