+++ codex-rs/tui/src/entertainment/shimmer_text.rs
