+++ codex-rs/tui/src/entertainment/test_support.rs
