+++ codex-rs/tui/src/lib.rs
@@ -44,7 +44,7 @@ mod clipboard_paste;
 mod color;
 pub mod custom_terminal;
 mod diff_render;
-mod entertainment_texts;
+mod entertainment;
 mod exec_cell;
 mod exec_command;
 mod external_editor;
@@ -69,7 +69,6 @@ mod resume_picker;
 mod selection_list;
 mod session_log;
 mod shimmer;
-mod shimmer_text;
 mod slash_command;
 mod status;
 mod status_indicator_shimmer;
@@ -77,7 +76,7 @@ mod status_indicator_widget;
 mod streaming;
 mod style;
 mod terminal_palette;
-pub mod test_support;
+pub use entertainment::test_support;
 mod text_formatting;
 mod tooltips;
 mod tui;
